+++ src/lib.rs
@@ -17,11 +17,12 @@ pub use paperclip_core::util;
 #[cfg(feature = "v2")]
 pub use paperclip_macros::api_v2_schema_struct as api_v2_schema;
 
-#[cfg(feature = "actix")]
+#[cfg(feature = "actix-base")]
 pub mod actix {
     //! Plugin types, traits and macros for actix-web framework.
 
     pub use paperclip_actix::{api_v2_errors, api_v2_operation, Apiv2Schema, Apiv2Security};
     pub use paperclip_actix::{web, App, Mountable, OpenApiExt};
+    pub use paperclip_core::v2::{AcceptedJson, CreatedJson, NoContent};
     pub use paperclip_core::v2::{OperationModifier, ResponderWrapper, ResponseWrapper};
 }